Moved skill components into a sub folder
